DEV-28 - delay rule sample code - more fixes

+++ js/cryptocorp.js
@@ -126,7 +126,7 @@ var CryptoCorp = new function () {
 	
   this.getParameters = function( value, asset, period, delay ) {
     var parameters = { 
-      "velocity_1": { "value": parseInt(value), "asset": asset, "period": parseInt(period), "limited_keys": [ KeyIndex.USER ] },
+      "velocity_1": { "value": parseFloat(value), "asset": asset, "period": parseInt(period), "limited_keys": [ KeyIndex.USER ] },
       "delay_1": parseInt(delay),      
     };
     return parameters;
